refactor: make useGetEpicodes and useGetChannels hook
