+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "vlive-frontend",
-  "version": "0.2.0",
+  "version": "0.2.2",
   "description": "A vlive project frontend",
   "author": "Team brosis <eunbines2@gmail.com>",
   "contributors": [
