+++ src/components/Preview.js
@@ -1,43 +1,22 @@
 import React from 'react';
 import styled from '@emotion/styled';
 import Image from 'next/image';
-import useSWR from 'swr';
-
-async function fetcher() {
-  const baseUrl = process.env.NEXT_PUBLIC_API_URL;
-
-  try {
-    const response = await fetch(`${baseUrl}/episode`, {
-      method: 'GET',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-    });
-    const { data } = await response.json();
-
-    return data;
-  } catch (err) {
-    console.error(err);
-  }
-}
+import useGetEpisodes from '../hooks/useGetEpisodes';
 
 export default function Preview() {
-  const { data: episodes } = useSWR('/episode', fetcher);
+  const { episodes } = useGetEpisodes();
 
   return (
     <PreviewContainer>
       {episodes?.map((episode) => {
         return (
           <div key={episode._id} className="episode">
-            {episode !== null && (
-              <Image
-                src="https://source.unsplash.com/random"
-                alt="Episode thumbnail"
-                width={100}
-                height={100}
-              />
-            )}
-            {episode.thumbnail}
+            <Image
+              src="https://source.unsplash.com/random"
+              alt="Episode thumbnail"
+              width={100}
+              height={100}
+            />
             <h3>{episode.title}</h3>
           </div>
         );
