+++ src/hooks/useGetChannels.js
@@ -1,0 +1,25 @@
+import useSWR from 'swr';
+
+async function fetcher() {
+  const baseUrl = process.env.NEXT_PUBLIC_API_URL;
+
+  try {
+    const response = await fetch(`${baseUrl}/channel`, {
+      method: 'GET',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+    });
+    const { data: channels } = await response.json();
+
+    return channels;
+  } catch (err) {
+    return err;
+  }
+}
+
+export default function useGetEpisodes() {
+  const { data: channels, error } = useSWR('/channel', fetcher);
+
+  return { channels, error };
+}
